[Cloud Security] fix rules flaky test suite (#212198)
## Summary

This PR fixes the flakiness of of rules page test suite -
https://github.com/elastic/kibana/issues/178413
Issues there were handled are:
1. Removing the set up of fleet server before every test which caused
flakiness.
![Screenshot 2025-02-24 at 9 54
40](https://github.com/user-attachments/assets/6913be9d-45aa-46fa-9923-671ed0d67f98)

2. Fixing race conditions.
3. Use retry.try function to re-execute in places where flakiness was
observed due to not waiting enough time before doing the action.

### Checklist

Reviewers should verify this PR satisfies this list as well.
- [x] [Flaky Test
Runner](https://ci-stats.kibana.dev/trigger_flaky_test_runner/1) was
used on any tests changed
- [x] The PR description includes the appropriate Release Notes section,
and the correct `release_note:*` label is applied per the
[guidelines](https://www.elastic.co/guide/en/kibana/master/contributing.html#kibana-release-notes-process)

(cherry picked from commit 14e7f6007e26bbdac23f65840747e49ba22581ea)
